+++ app/NotificationSetting.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App;
+
+use App\Models\User;
+use Auth;
+use Illuminate\Database\Eloquent\Model;
+
+class NotificationSetting extends Model
+{
+    protected $casts = [
+        'send_daily' => 'boolean',
+        'send_weekly' => 'boolean',
+    ];
+
+    protected $attributes = [
+        'send_daily' => false,
+        'send_weekly' => false,
+    ];
+
+    protected $fillable = [
+        'send_daily',
+        'send_weekly'
+    ];
+}