+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add log viewer
 - Add `date_of_birth` field (#3)
 - Add map with available addresses on the contact addresses detail view
+- Add notification settings
 
 ### Changed
 - Move `User` class into `Models` directory